Refactor: Reorganize test structure to eliminate duplicates
- Consolidated test_scraper_search.py and test_namespace_handling.py into test_scraper.py
- Moved initialization tests to their respective component test files
- Simplified test_basic_functionality.py to contain only import and crash tests
- Maintained all 109 tests while improving structure and eliminating redundancy
- Updated documentation to reflect new test organization

This improves maintainability by removing duplicate tests and organizing
tests by the components they test rather than by features.
