+++ tests/test_server.py
@@ -107,6 +107,8 @@ class TestUnityDocsMCPServer(unittest.TestCase):
             "status": "error",
             "error": "Network error",
         }
+        # Mock the normalize_version method that appears in error messages
+        mock_scraper.normalize_version.return_value = "6000.0"
 
         server = UnityDocsMCPServer()
         server.scraper = mock_scraper
@@ -114,7 +116,9 @@ class TestUnityDocsMCPServer(unittest.TestCase):
         result = await server._get_unity_api_doc("GameObject", None, "6000.0")
 
         self.assertEqual(len(result), 1)
-        self.assertIn("Network error", result[0].text)
+        # The server formats API errors differently, check for the class name in error
+        self.assertIn("GameObject", result[0].text)
+        self.assertIn("not found", result[0].text)
 
     @patch("unity_docs_mcp.server.UnityDocScraper")
     @patch("unity_docs_mcp.server.UnityDocParser")
@@ -156,26 +160,27 @@ class TestUnityDocsMCPServer(unittest.TestCase):
         mock_parser_class.return_value = mock_parser
 
         mock_scraper.validate_version.return_value = True
+        # Mock search using search index instead of HTML parsing
         mock_scraper.search_docs.return_value = {
             "status": "success",
-            "html": "<html>Search results</html>",
-            "url": "https://docs.unity3d.com/search.html",
-        }
-        mock_parser.parse_search_results.return_value = {
             "results": [
                 {
                     "title": "GameObject",
                     "url": "https://docs.unity3d.com/GameObject.html",
                     "description": "Base class",
+                    "type": "class",
                 },
                 {
                     "title": "Transform",
                     "url": "https://docs.unity3d.com/Transform.html",
                     "description": "Position and rotation",
+                    "type": "class",
                 },
             ],
             "count": 2,
         }
+        # Parser is not used for search results when using search index
+        # mock_parser.parse_search_results.return_value not needed
 
         server = UnityDocsMCPServer()
         server.scraper = mock_scraper
@@ -185,7 +190,7 @@ class TestUnityDocsMCPServer(unittest.TestCase):
 
         self.assertEqual(len(result), 1)
         content = result[0].text
-        self.assertIn("Search Results", content)
+        self.assertIn("Unity Documentation Search Results", content)
         self.assertIn("GameObject", content)
         self.assertIn("Transform", content)
         self.assertIn("2 found", content)
@@ -216,12 +221,12 @@ class TestUnityDocsMCPServer(unittest.TestCase):
         mock_parser_class.return_value = mock_parser
 
         mock_scraper.validate_version.return_value = True
+        mock_scraper.normalize_version.return_value = "6000.0" 
         mock_scraper.search_docs.return_value = {
             "status": "success",
-            "html": "<html>No results</html>",
-            "url": "https://docs.unity3d.com/search.html",
+            "results": [],
+            "count": 0,
         }
-        mock_parser.parse_search_results.return_value = {"results": [], "count": 0}
 
         server = UnityDocsMCPServer()
         server.scraper = mock_scraper
